Revert "replace wait-for with pg_isready health check"
This reverts commit 5f5d156a9296a21106e69da3805348f88f5732db.

Docker compose removed support for conditions in 'depends on' in version 3.
Services will start when a dependent on service has started, there is no longer a way to run a healtcheck and wait until they are ready.
Docker recommend using wait-for, which is what we were doing https://docs.docker.com/compose/startup-order/